Ignore prefix when comparing versions
When Microsoft changed their naming scheme starting with .NET 5, version
comparison broke for the .NET releases.  E.g., "netcoreapp2.1" and "net5.0"
were considered different products resulting in version comparison to throw
an exception.  The easiest "fix" was to ignore prefixes altogether.

An effect of this change is that more care is needed when crafting
configuration since a typoed prefix will be ignored.